+++ app/src/main/java/com/echoist/linkedout/presentation/TabletHomeScreen.kt
@@ -54,123 +54,14 @@ fun TabletHomeRoute(
     writingViewModel: WritingViewModel = hiltViewModel(),
     statusCode: Int
 ) {
-    val drawerState = rememberDrawerState(DrawerValue.Closed)
-    val scope = rememberCoroutineScope()
-    var selectedMenu by remember { mutableStateOf("Default") }
-    var isLogoutClicked by remember { mutableStateOf(false) }
-    val scrollState = rememberScrollState()
 
-    ModalNavigationDrawer(
-        drawerState = drawerState,
-        drawerContent = {
-            Row {
-                Column(
-                    Modifier
-                        .width(360.dp)
-                        .fillMaxHeight()
-                        .background(Color(0xFF121212))
-                        .verticalScroll(scrollState),
-                ) {
-                    MyProfile(item = viewModel.getMyInfo()) { }
-                    HorizontalDivider(thickness = 6.dp, color = Color(0xFF191919))
-                    MyLinkedOutBar()
-                    LineChartExample(essayCounts = viewModel.essayCount)
-                    Spacer(modifier = Modifier.height(10.dp))
-                    HorizontalDivider(thickness = 6.dp, color = Color(0xFF191919))
-                    ShopDrawerItem()
-                    HorizontalDivider(thickness = 6.dp, color = Color(0xFF191919))
-                    TabletDrawableItems("화면 설정", selectedMenu == "화면 설정") {
-                        selectedMenu = "화면 설정"
-                    }
-                    TabletDrawableItems("환경 설정", selectedMenu == "환경 설정") {
-                        selectedMenu = "환경 설정"
-                    }
-                    TabletDrawableItems("고객지원", selectedMenu == "고객지원") {
-                        selectedMenu = "고객지원"
-                    }
-                    TabletDrawableItems("업데이트 기록", selectedMenu == "업데이트 기록") {
-                        selectedMenu = "업데이트 기록"
-                    }
-
-                    LogoutBtn { isLogoutClicked = true } //todo logout 기능 만들기
-                }
-                Box(Modifier.fillMaxSize()) {
-                    when (selectedMenu) {
-                        "Default" -> {
-                            Box(
-                                Modifier
-                                    .fillMaxSize()
-                                    .alpha(0.0f)
-                            )
-                        }
-
-                        "화면 설정" -> {
-                            TabletThemeModeScreen {
-                                selectedMenu = "Default"
-                            }
-                        }
-
-                        "환경 설정" -> {
-                            TabletSettingRoute(navController = navController) {
-                                selectedMenu = "Default"
-                            }
-                        }
-
-                        "고객지원" -> {
-                            TabletSupportRoute(onCloseClick = {
-                                selectedMenu = "Default"
-                            }, onClickSupport = {
-                                selectedMenu = "링크드아웃 고객센터"
-                            })
-                        }
-
-                        "업데이트 기록" -> {
-                            TabletUpdateHistoryRoute {
-                                selectedMenu = "Default"
-                            }
-                        }
-
-                        "링크드아웃 고객센터" -> {
-                            TabletLinkedOutSupportRoute(onBackPressed = {
-                                selectedMenu = "고객지원"
-                            }, onClickInquiry = {
-                                selectedMenu = "1:1 문의하기"
-                            })
-                        }
-
-                        "1:1 문의하기" -> {
-                            TabletInquiryScreen {
-                                selectedMenu = "링크드아웃 고객센터"
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    ) {
-        Column {
-            Box(modifier = Modifier.fillMaxSize()) {
-                GlideImage(
-                    model = R.drawable.home_basic_tablet,
-                    contentDescription = "home_img",
-                    modifier = Modifier.fillMaxSize(),
-                    contentScale = ContentScale.FillWidth
-                )
-                TabletMainTopBar({
-                    scope.launch {
-                        drawerState.apply {
-                            selectedMenu = "Default"
-                            if (isClosed) open() else close()
-                        }
-                    }
-                }, { navController.navigate("NotificationPage") },
-                    viewModel.isExistUnreadAlerts
-                )
-                {
-                    viewModel.isFirstUser = true //튜토리얼 화면띄우기위함
-                }
-            }
-        }
+    Box(modifier = Modifier.fillMaxSize()) {
+        GlideImage(
+            model = R.drawable.home_basic_tablet,
+            contentDescription = "home_img",
+            modifier = Modifier.fillMaxSize(),
+            contentScale = ContentScale.FillWidth
+        )
     }
 }
